Refactoring, sidechain table, versionbits

+++ src/sidechain.h
@@ -48,7 +48,7 @@ struct SidechainDeposit {
 
 struct SidechainLD {
     int8_t nSidechain;
-    int8_t nPrevBlockRef;
+    int16_t nPrevBlockRef;
     uint256 hashCritical;
 
     bool operator==(const SidechainLD& a) const;
